+++ public/assets/js/faas/edit-machinery-functions.js
@@ -80,6 +80,6 @@ function formatStringToInteger(num){
         return 0
     }
     else {
-        return parseInt(num.replaceAll('%',''))
+        return parseFloat(num.replaceAll('%',''))
     }
 }
