+++ public/assets/js/rpt/edit-land-functions.js
@@ -525,6 +525,6 @@ function formatStringToInteger(num){
         return 0
     }
     else {
-        return parseInt(num.replaceAll('%',''))
+        return parseFloat(num.replaceAll('%',''))
     }
 }
