+++ public/assets/js/rpt/edit-machinery-functions.js
@@ -284,6 +284,6 @@ function formatStringToInteger(num){
         return 0
     }
     else {
-        return parseInt(num.replaceAll('%',''))
+        return parseFloat(num.replaceAll('%',''))
     }
 }
