+++ public/assets/js/treasury/create-treasury-other-functions.js
@@ -272,6 +272,6 @@ function formatStringToInteger(num){
         return 0
     }
     else {
-        return parseInt(num.replaceAll('%',''))
+        return parseFloat(num.replaceAll('%',''))
     }
 }
